refactoring
(cherry picked from commit 5fd99e9af79ac4da5046a596071864f313ff0f70)
