+++ adapter/golang/src/test/java/at/jku/isse/ecco/adapter/golang/data/TokenArtifactDataTest.java
@@ -1,4 +1,4 @@
-package at.jku.isse.ecco.adapter.golang;
+package at.jku.isse.ecco.adapter.golang.data;
 
 import at.jku.isse.ecco.artifact.ArtifactData;
 import org.junit.jupiter.api.Test;
